misc/HtmlTestSuite/results/test-0937.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml" lang="en">
<head><title>HTMLSelectMenuElement Test: labels</title>
<link href="mailto:iopopesc@microsoft.com" rel="author" title="Ionel Popescu"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

</head><body><label for="selectmenu1" id="label1">Label 1</label>
<selectmenu id="selectmenu1">
  <option>one</option>
  <option>two</option>
  <option>three</option>
  <option>four</option>
</selectmenu>
<label for="selectmenu1" id="label2">Label 2</label>
<label for="invalid-selectmenu1" id="label3">Label 3</label>

<script>

test(() =&gt; {
  let selectMenu = document.getElementById("selectmenu1");

  assert_true(selectMenu.labels instanceof NodeList);
  assert_equals(selectMenu.labels.length, 2);
  assert_equals(selectMenu.labels[0], document.getElementById("label1"));
  assert_equals(selectMenu.labels[1], document.getElementById("label2"));
}, "Validate selectmenu.labels");

</script>
</body></html>
